web/Tap Tap/requirements.txt
Flask==3.0.3
Werkzeug==3.0.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
